List only started profiles if no argument given
Invoking 'tmuss' and 'tmuss -l|--list' produces the same output. This is
duplicated behaviour. Furthermore, we are usually only interested in
started profiles, for example when we want to switch to another profile.
Cluttered output doesn't help in this case.

This change lists only started profiles if no argument is given
('tmuss') and lists all available profiles if the '-l|--list' option is
given ('tmuss -l|--list').